+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/relatorios/RelatorioPneuService.java
@@ -26,6 +26,28 @@ public class RelatorioPneuService {
     @NotNull
     private final ProLogExceptionHandler exceptionHandler = Injection.provideProLogExceptionHandler();
 
+    public void getKmRodadoPorPneuPorVidaCsv(@NotNull final OutputStream outputStream,
+                                             @NotNull final List<Long> codUnidades) {
+        try {
+            dao.getKmRodadoPorPneuPorVidaCsv(outputStream, codUnidades);
+        } catch (final Throwable throwable) {
+            Log.e(TAG, "Erro ao buscar o relatório de km percorrido por pneu por vida (CSV)", throwable);
+            throw new RuntimeException(throwable);
+        }
+    }
+
+    @NotNull
+    public Report getKmRodadoPorPneuPorVidaReport(@NotNull final List<Long> codUnidades) throws ProLogException {
+        try {
+            return dao.getKmRodadoPorPneuPorVidaReport(codUnidades);
+        } catch (final Throwable throwable) {
+            Log.e(TAG, "Erro ao buscar o relatório de km percorrido por pneu por vida (REPORT)", throwable);
+            throw exceptionHandler.map(
+                    throwable,
+                    "Erro ao gerar relatório, tente novamente");
+        }
+    }
+
     public void getAfericoesAvulsasCsv(@NotNull final OutputStream outputStream,
                                        @NotNull final List<Long> codUnidades,
                                        @NotNull final String dataInicial,