+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Change Log
 ==========
 
+<a name="v0.0.77"></a>
+## Version [v0.0.77](https://github.com/luizfp/PrologWeb/compare/v0.0.76...v0.0.77) (2018-10-19)
+
+### Features
+* Cria relatório para mostrar o KM rodado por pneu e por vida
+
+### Refactors
+* Altera insert do relato para salvar a versão do app que fez a requisição
+
+### Bug Fixes
+* Corrige busca das alternativas disponíveis para realização de um relato
+* Corrige método de busca de um relato pelo seu código
+* Corrige query de relatório da solicitação de folga
+* Corrige leitura da hrSaida e hrEntrada no import do mapa
+
 <a name="v0.0.76"></a>
 ## Version [v0.0.76](https://github.com/luizfp/PrologWeb/compare/v0.0.75...v0.0.76) (2018-10-02)